oracle-backend/requirements.txt
flask
requests
python-dotenv
multiversx-sdk
rich
python-dotenv